Added new subsurface class in utils
I seperated out the surface-based stuff into Utils.py. There is now a 'subsurface' class that does all the pycortex based stuff. This does all the donkey work
that is needed before defining a 'stimulus'. see test/develop/CF_progress.

I also inserted a command to pad the subsurfaces for each hemisphere so that they can be combined and will have the same shape as the design matrix.

The surfaces are removed after they are no longer needed (after computing the distance matrices).